requirements.txt
aiogram>=3.0
aiosqlite
aiohttp
beautifulsoup4
python-dotenv